+++ lendingservice-frontend/src/App.css
@@ -30,6 +30,7 @@
   padding: 5px;
   border-top: 2px solid;
   border-radius: 5px 5px 0 0;
+  text-decoration: none;
 }
 
 .tab:hover {
@@ -41,6 +42,8 @@
 }
 
 
+
+/* old stuff */
 @media (prefers-reduced-motion: no-preference) {
   .App-logo {
     animation: App-logo-spin infinite 20s linear;
