+++ lendingservice-frontend/src/ItemBrowse.tsx
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+function ItemBrowse() {
+    return (
+      <div className="BrowseItems">
+        <ul>
+          <li>Item 1<br />Address<br />City State Zip<br /></li>
+          <li>Item 2<br />Address<br />City State Zip<br /></li>
+          <li>Item 3<br />Address<br />City State Zip<br /></li>
+          <li>Item 4<br />Address<br />City State Zip<br /></li>
+        </ul>
+      </div>
+    )
+  }
+
+export default ItemBrowse;
