+++ lendingservice-frontend/src/MainNav.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+function MainNav() {
+    return(
+      <ul className="Nav">
+        <li>Nav Item 1</li>
+        <li>Nav Item 2</li>
+        <li>Nav Item 3</li>
+      </ul>
+    )
+  }
+
+export default MainNav;
